always behave as a single large keyboard
on the one hand this causes redundant scans of unconnected rows, but
the code gets simpler since it doesn't need to keep track which side it
is.
it also further simplifies the situation should the two halfs be
connected via serial. then we need to keep track of the whole state
anyway.