Don't disable the interrupts, just the interrupt lines
technically EXTI0_1_IRQn and EXTI2_3_IRQn could be disabled, but
EXTI4_15_IRQn is needed for pin A9 (the UART pin).

The interrupt on Pin A9 is used in completely different states that the
GPIO interrupts.

Pin A9 is used to detect when the other Keyboard half is able to receive
the state pressed buttons.

GPIOs interrupts are only used in the suspend state for power saving.

Both conditions could potentially overlap, so don't touch the shared
resource to avoid a conflict.