add rudimentary board to board communication
  - enable serial communication after 3s timeout
  - wait for other side to be ready
  - append crc checksum to message

Pull-Up is used to singal the other board that we are ready to receive.

Attention: Hardware falsely recognises a start bit if there isn't a long
enough waiting period that ensures the line is truy read as HIGH and
nothing in between.

After setting the pull-up, we wait until a minimum of 8 consecutive
reads on that pin return high.

TODO: cancel receive (break the aforementioned loop), if something pulls
down the line for too long.