implement remote wake via serial
I don't know why this was such hart nut to crack. But timers not
behaving as they should certainly didn't help.

TIM17 is used for transitioning from usb to uart
TIM15 is used to detect suspend over uart

I hope to be able to clean this up a little more and make it more
robust. If you touch the code, things just might stop working if the
state changes are touched just slightly.